throttle: enf: Move constants to one class
