+++ enforcer/src/main/java/org/wso2/micro/gateway/enforcer/throttle/ThrottleConstants.java
@@ -36,11 +36,31 @@ public class ThrottleConstants {
     public static final String THROTTLE_OUT_MESSAGE = "Message throttled out";
     public static final String THROTTLE_OUT_DESCRIPTION = "You have exceeded your quota";
     public static final String BLOCKING_MESSAGE = "Message blocked";
-    public static final String BLOCKING_DESCRIPTION = "You have been blocked from accesing the resource";
+    public static final String BLOCKING_DESCRIPTION = "You have been blocked from accessing the resource";
 
     public static final String THROTTLE_OUT_REASON_API_LIMIT_EXCEEDED = "API_LIMIT_EXCEEDED";
     public static final String THROTTLE_OUT_REASON_RESOURCE_LIMIT_EXCEEDED = "RESOURCE_LIMIT_EXCEEDED";
     public static final String THROTTLE_OUT_REASON_SUBSCRIPTION_LIMIT_EXCEEDED = "SUBSCRIPTION_LIMIT_EXCEEDED";
     public static final String THROTTLE_OUT_REASON_APPLICATION_LIMIT_EXCEEDED = "APPLICATION_LIMIT_EXCEEDED";
     public static final String POLICY_NOT_FOUND_DESCRIPTION = "POLICY ENFORCEMENT ERROR";
+
+    public static final String UNLIMITED_TIER = "Unlimited";
+    public static final String IP = "ip";
+    public static final String IPV6 = "ipv6";
+    public static final String BLOCKING_CONDITION_KEY = "blockingCondition";
+    public static final String POLICY_TEMPLATE_KEY = "keyTemplateValue";
+    public static final String THROTTLE_KEY = "throttleKey";
+    public static final String THROTTLE_OUT_REASON = "THROTTLED_OUT_REASON";
+    public static final String TOPIC_THROTTLE_DATA = "throttleData";
+
+    /**
+     * Advanced Throttling related constants.
+     */
+    public static class AdvancedThrottleConstants {
+        public static final String IS_THROTTLED = "isThrottled";
+        public static final String THROTTLE_KEY = "throttleKey";
+        public static final String EXPIRY_TIMESTAMP = "expiryTimeStamp";
+        public static final String EVALUATED_CONDITIONS = "evaluatedConditions";
+        public static final String TRUE = "true";
+    }
 }
