+++ enforcer/src/main/java/org/wso2/micro/gateway/enforcer/throttle/ThrottleEventListener.java
@@ -81,7 +81,7 @@ public class ThrottleEventListener implements MessageListener {
                 topicConnection = connFactory.createTopicConnection();
                 topicConnection.start();
                 topicSession = topicConnection.createTopicSession(false, TopicSession.AUTO_ACKNOWLEDGE);
-                Topic gatewayJmsTopic = topicSession.createTopic("throttleData");
+                Topic gatewayJmsTopic = topicSession.createTopic(ThrottleConstants.TOPIC_THROTTLE_DATA);
                 TopicSubscriber listener = topicSession.createSubscriber(gatewayJmsTopic);
                 listener.setMessageListener(new ThrottleEventListener());
             } catch (NamingException | JMSException e) {
@@ -112,8 +112,8 @@ public class ThrottleEventListener implements MessageListener {
                 map.put(key, mapMessage.getObject(key));
             }
 
-            if (APIConstants.TopicNames.TOPIC_THROTTLE_DATA.equalsIgnoreCase(jmsDestination.getTopicName())) {
-                if (map.get(APIConstants.THROTTLE_KEY) != null) {
+            if (ThrottleConstants.TOPIC_THROTTLE_DATA.equalsIgnoreCase(jmsDestination.getTopicName())) {
+                if (map.get(ThrottleConstants.THROTTLE_KEY) != null) {
                     /*
                      * This message contains throttle data in map which contains Keys
                      * throttleKey - Key of particular throttling level
@@ -130,10 +130,11 @@ public class ThrottleEventListener implements MessageListener {
     }
 
     private void handleThrottleUpdateMessage(Map<String, Object> map) {
-        String throttleKey = map.get(APIConstants.AdvancedThrottleConstants.THROTTLE_KEY).toString();
-        String throttleState = map.get(APIConstants.AdvancedThrottleConstants.IS_THROTTLED).toString();
-        long timeStamp = Long.parseLong(map.get(APIConstants.AdvancedThrottleConstants.EXPIRY_TIMESTAMP).toString());
-        Object evaluatedConditionObject = map.get(APIConstants.AdvancedThrottleConstants.EVALUATED_CONDITIONS);
+        String throttleKey = map.get(ThrottleConstants.AdvancedThrottleConstants.THROTTLE_KEY).toString();
+        String throttleState = map.get(ThrottleConstants.AdvancedThrottleConstants.IS_THROTTLED).toString();
+        long timeStamp = Long.parseLong(map.get(ThrottleConstants.
+                AdvancedThrottleConstants.EXPIRY_TIMESTAMP).toString());
+        Object evaluatedConditionObject = map.get(ThrottleConstants.AdvancedThrottleConstants.EVALUATED_CONDITIONS);
         ThrottleDataHolder dataHolder = ThrottleDataHolder.getInstance();
 
         if (log.isDebugEnabled()) {
@@ -141,8 +142,7 @@ public class ThrottleEventListener implements MessageListener {
                     throttleState + ", expiryTime: " + new Date(timeStamp).toString());
         }
 
-        if (APIConstants.AdvancedThrottleConstants.TRUE.equalsIgnoreCase(throttleState)) {
-
+        if (ThrottleConstants.AdvancedThrottleConstants.TRUE.equalsIgnoreCase(throttleState)) {
             dataHolder.addThrottleData(throttleKey, timeStamp);
             APICondition extractedKey = extractAPIorResourceKey(throttleKey);
             log.debug("Adding throttling key : {}",  extractedKey);
